Move server.session.* to server.servlet.session.*
Closes gh-11589
