+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/web/ServerProperties.java
@@ -98,8 +98,6 @@ public class ServerProperties {
 	 */
 	private Duration connectionTimeout;
 
-	private final Session session = new Session();
-
 	@NestedConfigurationProperty
 	private Ssl ssl;
 
@@ -177,10 +175,6 @@ public class ServerProperties {
 		return this.error;
 	}
 
-	public Session getSession() {
-		return this.session;
-	}
-
 	public Ssl getSsl() {
 		return this.ssl;
 	}
@@ -236,6 +230,8 @@ public class ServerProperties {
 		@NestedConfigurationProperty
 		private final Jsp jsp = new Jsp();
 
+		private final Session session = new Session();
+
 		public String getContextPath() {
 			return this.contextPath;
 		}
@@ -266,6 +262,10 @@ public class ServerProperties {
 
 		public Jsp getJsp() {
 			return this.jsp;
+		}
+
+		public Session getSession() {
+			return this.session;
 		}
 
 		public String getServletMapping() {
@@ -319,195 +319,195 @@ public class ServerProperties {
 			return result;
 		}
 
-	}
-
-	/**
-	 * Session properties.
-	 */
-	public static class Session {
-
 		/**
-		 * Session timeout. If a duration suffix is not specified, seconds will be used.
+		 * Session properties.
 		 */
-		@DefaultDurationUnit(ChronoUnit.SECONDS)
-		private Duration timeout;
-
-		/**
-		 * Session tracking modes (one or more of the following: "cookie", "url", "ssl").
-		 */
-		private Set<SessionTrackingMode> trackingModes;
-
-		/**
-		 * Whether to persist session data between restarts.
-		 */
-		private boolean persistent;
-
-		/**
-		 * Directory used to store session data.
-		 */
-		private File storeDir;
-
-		private final Cookie cookie = new Cookie();
-
-		public Cookie getCookie() {
-			return this.cookie;
-		}
-
-		public Duration getTimeout() {
-			return this.timeout;
-		}
-
-		public void setTimeout(Duration timeout) {
-			this.timeout = timeout;
-		}
-
-		public Set<SessionTrackingMode> getTrackingModes() {
-			return this.trackingModes;
-		}
-
-		public void setTrackingModes(Set<SessionTrackingMode> trackingModes) {
-			this.trackingModes = trackingModes;
-		}
-
-		public boolean isPersistent() {
-			return this.persistent;
-		}
-
-		public void setPersistent(boolean persistent) {
-			this.persistent = persistent;
-		}
-
-		public File getStoreDir() {
-			return this.storeDir;
-		}
-
-		public void setStoreDir(File storeDir) {
-			this.storeDir = storeDir;
-		}
-
-		/**
-		 * Cookie properties.
-		 */
-		public static class Cookie {
+		public static class Session {
 
 			/**
-			 * Session cookie name.
-			 */
-			private String name;
-
-			/**
-			 * Domain for the session cookie.
-			 */
-			private String domain;
-
-			/**
-			 * Path of the session cookie.
-			 */
-			private String path;
-
-			/**
-			 * Comment for the session cookie.
-			 */
-			private String comment;
-
-			/**
-			 * "HttpOnly" flag for the session cookie.
-			 */
-			private Boolean httpOnly;
-
-			/**
-			 * "Secure" flag for the session cookie.
-			 */
-			private Boolean secure;
-
-			/**
-			 * Maximum age of the session cookie.
+			 * Session timeout. If a duration suffix is not specified, seconds will be used.
 			 */
 			@DefaultDurationUnit(ChronoUnit.SECONDS)
-			private Duration maxAge;
+			private Duration timeout;
 
-			public String getName() {
-				return this.name;
+			/**
+			 * Session tracking modes (one or more of the following: "cookie", "url", "ssl").
+			 */
+			private Set<SessionTrackingMode> trackingModes;
+
+			/**
+			 * Whether to persist session data between restarts.
+			 */
+			private boolean persistent;
+
+			/**
+			 * Directory used to store session data.
+			 */
+			private File storeDir;
+
+			private final Cookie cookie = new Cookie();
+
+			public Cookie getCookie() {
+				return this.cookie;
 			}
 
-			public void setName(String name) {
-				this.name = name;
+			public Duration getTimeout() {
+				return this.timeout;
 			}
 
-			public String getDomain() {
-				return this.domain;
+			public void setTimeout(Duration timeout) {
+				this.timeout = timeout;
 			}
 
-			public void setDomain(String domain) {
-				this.domain = domain;
+			public Set<SessionTrackingMode> getTrackingModes() {
+				return this.trackingModes;
 			}
 
-			public String getPath() {
-				return this.path;
+			public void setTrackingModes(Set<SessionTrackingMode> trackingModes) {
+				this.trackingModes = trackingModes;
 			}
 
-			public void setPath(String path) {
-				this.path = path;
+			public boolean isPersistent() {
+				return this.persistent;
 			}
 
-			public String getComment() {
-				return this.comment;
+			public void setPersistent(boolean persistent) {
+				this.persistent = persistent;
 			}
 
-			public void setComment(String comment) {
-				this.comment = comment;
+			public File getStoreDir() {
+				return this.storeDir;
 			}
 
-			public Boolean getHttpOnly() {
-				return this.httpOnly;
+			public void setStoreDir(File storeDir) {
+				this.storeDir = storeDir;
 			}
 
-			public void setHttpOnly(Boolean httpOnly) {
-				this.httpOnly = httpOnly;
+			/**
+			 * Cookie properties.
+			 */
+			public static class Cookie {
+
+				/**
+				 * Session cookie name.
+				 */
+				private String name;
+
+				/**
+				 * Domain for the session cookie.
+				 */
+				private String domain;
+
+				/**
+				 * Path of the session cookie.
+				 */
+				private String path;
+
+				/**
+				 * Comment for the session cookie.
+				 */
+				private String comment;
+
+				/**
+				 * "HttpOnly" flag for the session cookie.
+				 */
+				private Boolean httpOnly;
+
+				/**
+				 * "Secure" flag for the session cookie.
+				 */
+				private Boolean secure;
+
+				/**
+				 * Maximum age of the session cookie.
+				 */
+				@DefaultDurationUnit(ChronoUnit.SECONDS)
+				private Duration maxAge;
+
+				public String getName() {
+					return this.name;
+				}
+
+				public void setName(String name) {
+					this.name = name;
+				}
+
+				public String getDomain() {
+					return this.domain;
+				}
+
+				public void setDomain(String domain) {
+					this.domain = domain;
+				}
+
+				public String getPath() {
+					return this.path;
+				}
+
+				public void setPath(String path) {
+					this.path = path;
+				}
+
+				public String getComment() {
+					return this.comment;
+				}
+
+				public void setComment(String comment) {
+					this.comment = comment;
+				}
+
+				public Boolean getHttpOnly() {
+					return this.httpOnly;
+				}
+
+				public void setHttpOnly(Boolean httpOnly) {
+					this.httpOnly = httpOnly;
+				}
+
+				public Boolean getSecure() {
+					return this.secure;
+				}
+
+				public void setSecure(Boolean secure) {
+					this.secure = secure;
+				}
+
+				public Duration getMaxAge() {
+					return this.maxAge;
+				}
+
+				public void setMaxAge(Duration maxAge) {
+					this.maxAge = maxAge;
+				}
+
 			}
 
-			public Boolean getSecure() {
-				return this.secure;
-			}
+			/**
+			 * Available session tracking modes (mirrors
+			 * {@link javax.servlet.SessionTrackingMode}.
+			 */
+			public enum SessionTrackingMode {
 
-			public void setSecure(Boolean secure) {
-				this.secure = secure;
-			}
+				/**
+				 * Send a cookie in response to the client's first request.
+				 */
+				COOKIE,
 
-			public Duration getMaxAge() {
-				return this.maxAge;
-			}
+				/**
+				 * Rewrite the URL to append a session ID.
+				 */
+				URL,
 
-			public void setMaxAge(Duration maxAge) {
-				this.maxAge = maxAge;
+				/**
+				 * Use SSL build-in mechanism to track the session.
+				 */
+				SSL
+
 			}
 
 		}
-
-		/**
-		 * Available session tracking modes (mirrors
-		 * {@link javax.servlet.SessionTrackingMode}.
-		 */
-		public enum SessionTrackingMode {
-
-			/**
-			 * Send a cookie in response to the client's first request.
-			 */
-			COOKIE,
-
-			/**
-			 * Rewrite the URL to append a session ID.
-			 */
-			URL,
-
-			/**
-			 * Use SSL build-in mechanism to track the session.
-			 */
-			SSL
-
-		}
-
 	}
+
 
 	/**
 	 * Tomcat properties.
