+++ spring-boot-project/spring-boot-autoconfigure/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1021,6 +1021,96 @@
       }
     },
     {
+      "name" : "server.session.cookie.comment",
+      "type" : "java.lang.String",
+      "description" : "Comment for the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.comment",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.domain",
+      "type" : "java.lang.String",
+      "description" : "Domain for the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.domain",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.http-only",
+      "type" : "java.lang.Boolean",
+      "description" : "\"HttpOnly\" flag for the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.http-only",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.max-age",
+      "type" : "java.time.Duration",
+      "description" : "Maximum age of the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.max-age",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.name",
+      "type" : "java.lang.String",
+      "description" : "Session cookie name.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.name",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.path",
+      "type" : "java.lang.String",
+      "description" : "Path of the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.path",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.cookie.secure",
+      "type" : "java.lang.Boolean",
+      "description" : "\"Secure\" flag for the session cookie.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.cookie.secure",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.persistent",
+      "type" : "java.lang.Boolean",
+      "description" : "Whether to persist session data between restarts.",
+      "defaultValue" : false,
+      "deprecation" : {
+        "replacement" : "server.servlet.session.persistent",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.store-dir",
+      "type" : "java.io.File",
+      "description" : "Directory used to store session data.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.store-dir",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.timeout",
+      "type" : "java.time.Duration",
+      "description" : "Session timeout. If a duration suffix is not specified, seconds will be used.",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.timeout",
+        "level" : "error"
+      }
+    }, {
+      "name" : "server.session.tracking-modes",
+      "type" : "java.util.Set<org.springframework.boot.autoconfigure.web.ServerProperties.Session.SessionTrackingMode>",
+      "description" : "Session tracking modes (one or more of the following: \"cookie\", \"url\", \"ssl\").",
+      "deprecation" : {
+        "replacement" : "server.servlet.session.tracking-modes",
+        "level" : "error"
+      }
+    },
+    {
       "name": "server.undertow.buffers-per-region",
       "type": "java.lang.Integer",
       "description": "Number of buffer per region.",
