+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/web/servlet/DefaultServletWebServerFactoryCustomizerTests.java
@@ -212,15 +212,15 @@ public class DefaultServletWebServerFactoryCustomizerTests {
 	@Test
 	public void customizeSessionProperties() throws Exception {
 		Map<String, String> map = new HashMap<>();
-		map.put("server.session.timeout", "123");
-		map.put("server.session.tracking-modes", "cookie,url");
-		map.put("server.session.cookie.name", "testname");
-		map.put("server.session.cookie.domain", "testdomain");
-		map.put("server.session.cookie.path", "/testpath");
-		map.put("server.session.cookie.comment", "testcomment");
-		map.put("server.session.cookie.http-only", "true");
-		map.put("server.session.cookie.secure", "true");
-		map.put("server.session.cookie.max-age", "60");
+		map.put("server.servlet.session.timeout", "123");
+		map.put("server.servlet.session.tracking-modes", "cookie,url");
+		map.put("server.servlet.session.cookie.name", "testname");
+		map.put("server.servlet.session.cookie.domain", "testdomain");
+		map.put("server.servlet.session.cookie.path", "/testpath");
+		map.put("server.servlet.session.cookie.comment", "testcomment");
+		map.put("server.servlet.session.cookie.http-only", "true");
+		map.put("server.servlet.session.cookie.secure", "true");
+		map.put("server.servlet.session.cookie.max-age", "60");
 		bindProperties(map);
 		ConfigurableServletWebServerFactory factory = mock(
 				ConfigurableServletWebServerFactory.class);
@@ -536,7 +536,7 @@ public class DefaultServletWebServerFactoryCustomizerTests {
 	@Test
 	public void sessionStoreDir() {
 		Map<String, String> map = new HashMap<>();
-		map.put("server.session.store-dir", "myfolder");
+		map.put("server.servlet.session.store-dir", "myfolder");
 		bindProperties(map);
 		JettyServletWebServerFactory factory = spy(new JettyServletWebServerFactory());
 		this.customizer.customize(factory);
