+++ spring-boot-project/spring-boot-docs/src/main/asciidoc/appendix-application-properties.adoc
@@ -194,17 +194,17 @@ content into your application. Rather, pick only the properties that you need.
 	server.servlet.jsp.init-parameters.*= # Init parameters used to configure the JSP servlet.
 	server.servlet.jsp.registered=true # Whether the JSP servlet is registered.
 	server.servlet.path=/ # Path of the main dispatcher servlet.
-	server.session.cookie.comment= # Comment for the session cookie.
-	server.session.cookie.domain= # Domain for the session cookie.
-	server.session.cookie.http-only= # "HttpOnly" flag for the session cookie.
-	server.session.cookie.max-age= # Maximum age of the session cookie. If a duration suffix is not specified, seconds will be used.
-	server.session.cookie.name= # Session cookie name.
-	server.session.cookie.path= # Path of the session cookie.
-	server.session.cookie.secure= # "Secure" flag for the session cookie.
-	server.session.persistent=false # Whether to persist session data between restarts.
-	server.session.store-dir= # Directory used to store session data.
-	server.session.timeout= # Session timeout. If a duration suffix is not specified, seconds will be used.
-	server.session.tracking-modes= # Session tracking modes (one or more of the following: "cookie", "url", "ssl").
+	server.servlet.session.cookie.comment= # Comment for the session cookie.
+	server.servlet.session.cookie.domain= # Domain for the session cookie.
+	server.servlet.session.cookie.http-only= # "HttpOnly" flag for the session cookie.
+	server.servlet.session.cookie.max-age= # Maximum age of the session cookie. If a duration suffix is not specified, seconds will be used.
+	server.servlet.session.cookie.name= # Session cookie name.
+	server.servlet.session.cookie.path= # Path of the session cookie.
+	server.servlet.session.cookie.secure= # "Secure" flag for the session cookie.
+	server.servlet.session.persistent=false # Whether to persist session data between restarts.
+	server.servlet.session.store-dir= # Directory used to store session data.
+	server.servlet.session.timeout= # Session timeout. If a duration suffix is not specified, seconds will be used.
+	server.servlet.session.tracking-modes= # Session tracking modes (one or more of the following: "cookie", "url", "ssl").
 	server.ssl.ciphers= # Supported SSL ciphers.
 	server.ssl.client-auth= # Whether client authentication is wanted ("want") or needed ("need"). Requires a trust store.
 	server.ssl.enabled= # Enable SSL support.
